requirements.txt
clize>=5.0.2
google-genai>=1.7.0
matplotlib>=3.9.4
numpy>=2.0.2
openai>=1.63.0
pydantic>=2.10.6
python-dotenv>=1.0.1
python_docx>=1.1.2
rich>=13.9.4
soundfile>=0.13.1
text2qti>=0.7.1
tomli>=2.2.1
